move to a template to compute PF isolation for both EGamma and Muons
