+++ RecoEgamma/EgammaHLTProducers/src/SealModules.cc
@@ -39,8 +39,8 @@
 #include "RecoEgamma/EgammaHLTProducers/interface/EgammaHLTPFChargedIsolationProducer.h"
 #include "RecoEgamma/EgammaHLTProducers/interface/EgammaHLTPFNeutralIsolationProducer.h"
 
-#include "RecoEgamma/EgammaHLTProducers/interface/EgammaHLTEcalPFClusterIsolationProducer.h"
-#include "RecoEgamma/EgammaHLTProducers/interface/EgammaHLTHcalPFClusterIsolationProducer.h"
+//#include "RecoEgamma/EgammaHLTProducers/interface/EgammaHLTEcalPFClusterIsolationProducer.h"
+//#include "RecoEgamma/EgammaHLTProducers/interface/EgammaHLTHcalPFClusterIsolationProducer.h"
 
 #include "RecoEgamma/EgammaHLTProducers/interface/EgammaHLTCaloTowerProducer.h"
 
@@ -79,8 +79,8 @@ DEFINE_FWK_MODULE(EgammaHLTPFPhotonIsolationProducer);
 DEFINE_FWK_MODULE(EgammaHLTPFChargedIsolationProducer);
 DEFINE_FWK_MODULE(EgammaHLTPFNeutralIsolationProducer);
 
-DEFINE_FWK_MODULE(EgammaHLTEcalPFClusterIsolationProducer);
-DEFINE_FWK_MODULE(EgammaHLTHcalPFClusterIsolationProducer);
+//DEFINE_FWK_MODULE(EgammaHLTEcalPFClusterIsolationProducer);
+//DEFINE_FWK_MODULE(EgammaHLTHcalPFClusterIsolationProducer);
 
 DEFINE_FWK_MODULE(EgammaHLTCaloTowerProducer);
